Jacobi solver with SOR/SUR added

+++ Compact Schemes/poisson_solver.jl
@@ -27,13 +27,13 @@ flag_solver             = Int32(input_parameters[3])
 flag_multigrid          = Int32(input_parameters[4])
 flag_start              = Int32(input_parameters[5])
 flag_restriction        = Int32(input_parameters[6])
-tolerance               = Float32(input_parameters[7])
-omega                   = Float32(input_parameters[8])
+tolerance               = Float64(input_parameters[7])
+omega                   = Float64(input_parameters[8])
 relaxation_f2c          = Int32(input_parameters[9])
 relaxation_c2f          = Int32(input_parameters[10])
 relaxation_coarsest     = Int32(input_parameters[11])
 maximum_iterations      = Int32(input_parameters[12])
-tiny                    = Float32(input_parameters[13])
+tiny                    = Float64(input_parameters[13])
 k = 1
 
 # Assign the domain size based on initial problem
@@ -56,11 +56,11 @@ dx = (x_right - x_left)/nx
 dy = (y_top - y_bottom)/ny
 
 # allocate array for x and y position of grids, exact solution and source term
-x_position      = Array{Float32}(undef, nx+1)
-y_position      = Array{Float32}(undef, ny+1)
-u_exact         = Array{Float32}(undef, nx+1, ny+1)
-source          = Array{Float32}(undef, nx+1, ny+1)
-u_numerical     = Array{Float32}(undef, nx+1, ny+1)
+x_position      = Array{Float64}(undef, nx+1)
+y_position      = Array{Float64}(undef, ny+1)
+u_exact         = Array{Float64}(undef, nx+1, ny+1)
+source          = Array{Float64}(undef, nx+1, ny+1)
+u_numerical     = Array{Float64}(undef, nx+1, ny+1)
 
 # assign the computational domain, exact solution, source term ,
 # initial condition, boundary condition
@@ -74,7 +74,7 @@ initial_condition(u_numerical, nx, ny, flag_start)
 boundary_condition(u_numerical, u_exact, nx, ny)
 
 # calculate residual and initial l2 norm of the residual
-residual         = zeros(Float32, nx+1, ny+1)
+residual         = zeros(Float64, nx+1, ny+1)
 initial_rms      = 0.0
 rms              = 0.0
 compute_residual(nx, ny, dx, dy, source, u_numerical, residual)
